WIP Team Synergy addition
Logic to parse two xml files (need to unify). Mainframe knows how to
store synergy data, calculate synergy of a team, and do a simple
weighting of synergies and versus-scores to size up two rosters against
each other. Deeper predictions have yet to be put in place. (The runtime
behavior of the Mainframe is not yet affected.)